Reimplemented Data to fix bugs
The old implementation made implicit assumptions about the object not
being copied: it stored a pointer to a member in another member. If
the object was copied later on, the pointer would be invalid. I'm
pretty sure that's what was causing the failures on Ubuntu.

I reimplemented it to avoid this. It's kind of messy, and there may be
simpler ways to do it, but let's see if this works for now...

#3